+++ library/src/types/other.ts
@@ -1,14 +1,3 @@
-import type {
-  LazySchema,
-  LazySchemaAsync,
-  NonNullableIssue,
-  NonNullableSchema,
-  NonNullableSchemaAsync,
-  NullishSchema,
-  NullishSchemaAsync,
-  OptionalSchema,
-  OptionalSchemaAsync,
-} from '../schemas/index.ts';
 import type { Config } from './config.ts';
 import type { Dataset } from './dataset.ts';
 import type { InferInput, InferIssue } from './infer.ts';
@@ -79,43 +68,3 @@ export type DefaultValue<
       ? Awaited<ReturnType<TDefault>>
       : TDefault
     : never;
-
-/**
- * Question mark schema type.
- *
- * TODO: Document that for simplicity and bundle size, we currently do not
- * distinguish between `undefined` and missing keys when using `optional` and
- * `nullish`.
- */
-export type QuestionMarkSchema =
-  | NullishSchema<BaseSchema<unknown, unknown, BaseIssue<unknown>>, unknown>
-  | OptionalSchema<BaseSchema<unknown, unknown, BaseIssue<unknown>>, unknown>
-  // @ts-expect-error
-  | LazySchema<QuestionMarkSchema>
-  | NonNullableSchema<
-      // @ts-expect-error
-      QuestionMarkSchema,
-      ErrorMessage<NonNullableIssue> | undefined
-    >;
-
-/**
- * Question mark schema async type.
- */
-export type QuestionMarkSchemaAsync =
-  | NullishSchemaAsync<
-      | BaseSchema<unknown, unknown, BaseIssue<unknown>>
-      | BaseSchemaAsync<unknown, unknown, BaseIssue<unknown>>,
-      unknown
-    >
-  | OptionalSchemaAsync<
-      | BaseSchema<unknown, unknown, BaseIssue<unknown>>
-      | BaseSchemaAsync<unknown, unknown, BaseIssue<unknown>>,
-      unknown
-    >
-  // @ts-expect-error
-  | LazySchemaAsync<QuestionMarkSchema | QuestionMarkSchemaAsync>
-  | NonNullableSchemaAsync<
-      // @ts-expect-error
-      QuestionMarkSchema | QuestionMarkSchemaAsync,
-      ErrorMessage<NonNullableIssue> | undefined
-    >;