+++ website/src/components/Credits.tsx
@@ -84,13 +84,11 @@ export const Credits = component$(() => {
       <ul class="mt-4 flex flex-wrap gap-2 md:mt-5 lg:mt-6 lg:gap-3">
         {[
           'Thanaen',
-          'KATT',
           'osdiab',
           'ruiaraujo012',
           'hyunbinseo',
           'F0rce',
           'caegdeveloper',
-          'luckasnix',
         ].map((sponsor) => (
           <li key={sponsor}>
             <a