+++ library/CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to the library will be documented in this file.
 ## vX.X.X (Month DD, YYYY)
 
 - Change `expects` and `expected` property by enclosing combined values in parentheses
+- Change question mark handling for `optional`, `optionalAsync`, `nullish` and `nullishAsync` schemas in objects
+- Fix TypeScript errors in `TuplePath` and `QuestionMarkSchema` type (issue #659, #776)
 
 ## v0.37.0 (July 30, 2024)